make the cache thread safe
there are several conditions where we have race conditions with cleaning
the shared library cache:
- when multiple jobs start at the same time and try to use an expired
  cached shared library, we run into troubles. The first one will start
  deleting the cache dir. The second one will see that the cachedir was
  modified and consider it uptodate and start copying from there.
- when mutiple jobs start at the same time with no cache entry available
  second will try to copy while the first is still cloning.
- the background cleaner is about to clean when a new job starts
- an administrator decides to clear the cache right before a new build
  wants to use the cache library

+++ src/main/java/org/jenkinsci/plugins/workflow/libs/LibraryCachingCleanup.java
@@ -1,12 +1,13 @@
 package org.jenkinsci.plugins.workflow.libs;
 
+import java.io.IOException;
+import java.util.concurrent.TimeUnit;
+import java.util.concurrent.locks.ReentrantReadWriteLock;
+
 import hudson.Extension;
 import hudson.FilePath;
 import hudson.model.AsyncPeriodicWork;
 import hudson.model.TaskListener;
-
-import java.io.IOException;
-import java.util.concurrent.TimeUnit;
 
 @Extension public class LibraryCachingCleanup extends AsyncPeriodicWork {
     private final Long recurrencePeriod;
@@ -27,8 +28,15 @@ import java.util.concurrent.TimeUnit;
         for (FilePath library: globalCacheDir.list()) {
             for (FilePath version: library.list()) {
                 final FilePath lastReadFile = new FilePath(version, LibraryCachingConfiguration.LAST_READ_FILE);
-                if (lastReadFile.exists() && (lastReadFile.lastModified() + unreadCacheClearTime) < System.currentTimeMillis()) {
-                    version.deleteRecursive();
+                ReentrantReadWriteLock retrieveLock = LibraryAdder.cacheRetrieveLock.get(version.getName());
+                if (retrieveLock.writeLock().tryLock()) {
+                    try {
+                      if (lastReadFile.exists() && (lastReadFile.lastModified() + unreadCacheClearTime) < System.currentTimeMillis()) {
+                          version.deleteRecursive();
+                      }
+                    } finally {
+                        retrieveLock.writeLock().unlock();
+                    }
                 }
             }
         }
